Start refactoring for v0.2.0 [no ci]
