+++ src/error.jl
@@ -4,16 +4,16 @@
 This error indicates any failure during QUBO formulation
 """
 struct QUBOError <: Exception
-    msg::Union{Nothing, String}
+    msg::Union{String, Nothing}
 
     function QUBOError(msg::Union{Nothing, String} = nothing)
-        return new(msg)
+        new(msg)
     end
 end
 
 function Base.showerror(io::IO, e::QUBOError)
     if isnothing(e.msg)
-        print(io, "The current model could not be converted to QUBO")
+        print(io, "The current model can't be converted to QUBO")
     else
         print(io, e.msg)
     end
