+++ src/virtual/VirtualMapping.jl
@@ -1,0 +1,15 @@
+module VirtualMapping
+
+# :: Pseudo-Boolean Optimization ::
+import ..PBO
+
+# -*- :: MathOptInterface :: -*-
+using MathOptInterface
+const MOI = MathOptInterface
+const VI = MOI.VariableIndex
+
+include("variable.jl")
+include("model.jl")
+include("encoding.jl")
+
+end
