+++ test/runtests.jl
@@ -7,28 +7,26 @@ const MOIU = MOI.Utilities
 const VI = MOI.VariableIndex
 
 # -*- Imports -*-
-using JuMP
+# using JuMP
 using ToQUBO
-using Anneal
+# using Anneal
 using LinearAlgebra
 using TOML
-
-const VM = ToQUBO.VirtualMapping
 
 # -*- Tests: Version -*-
 include("version.jl")
 
 # -*- Tests: Library -*-
-include("./lib/pbo.jl")
-include("./lib/virtual.jl")
+include(joinpath("lib", "pbo.jl"))
+# include(joinpath("lib", "virtual.jl"))
 
 # -*- Tests: Interface -*-
-# include("./lib/moi.jl")
-# include("./lib/jump.jl")
+# include(joinpath("interface", "moi.jl"))
+# include(joinpath("interface", "jump.jl"))
 
 # -*- Tests: Examples -*-
-@testset "Quantum Bridge Analytics I" begin
-    include("examples/qba2.jl")
-    include("examples/qba3_1.jl")
-    include("examples/qba3_2.jl")
-end
+# @testset "Quantum Bridge Analytics I" begin
+#     include(joinpath("examples", "qba2.jl"))
+#     include(joinpath("examples", "qba3_1.jl"))
+#     include(joinpath("examples", "qba3_2.jl"))
+# end
